Add false_safety/collapse_criticality failure-lattice rungs, wire safe_to_capture, persist tactical_events
Extended the piece-safety failure lattice (hanging_piece -> outnumbered ->
overloaded/false_defense) with the rung above both:

- false_safety: fires when the raw defender count alone says a piece is
  safe (raw_defender_count >= attacker_count) but isn't, once defenders
  pinned off-line or overloaded elsewhere are discounted - the case where
  a player "counted correctly" and still misses the tactic. New concept,
  ELO 1800, confidence 0.8; three new tests including a deliberately
  partial-compromise position that doesn't overlap with false_defense's
  all-or-nothing check.
- GatedIssue.stage: a machine-legible ladder-rung marker (1..4) alongside
  the existing phrase text, computed once via ladder_stage() and reused in
  both rank_issues_for_position/rank_issues_for_player.
- Rewrote outnumbered's phrase to state the shallower check that passed
  ("has a defender, so it isn't simply hanging, but...") matching
  false_defense/false_safety's existing "looks X but Y" framing.

Built ThreatGraph::collapse_criticality: clear the whole local cluster
contesting a square (every attacker, defender, occupant), then place each
candidate back individually against that clean board and read what changes
- no capture order, no recapture choice, just "if this piece is the one
left standing here, is that safe for it." Generalizes what
find_overloaded/find_false_defense each check narrowly (a pin is "placing
this piece here leaves its own king in check"; an overload is "placing it
here swings control of some other square") into one clear-and-place
mechanism. Went through two rejected designs first (least-valuable-attacker
substitution, then single-piece removal from an otherwise-cluttered board)
before landing on the clear-the-whole-cluster mechanic - both wrong turns
and why they were wrong are logged in PLAN.md.

Per-candidate output is fully identifiable, not just a verdict:
king-zone control deltas, ThreatGraph::checkers() naming which piece
actually delivers a check (own_king_checked_by/delivers_check_via, not
just a bool - is_in_check itself now defined in terms of checkers, one
computation not two), best-effort checkmate via a real shakmaty
Chess::from_setup construction (the one thing the graph genuinely can't
answer itself), and newly_hanging (reusing find_hanging on the
hypothetical to catch pieces elsewhere on the board that lose their
defender as a side effect).

Wired straight into the lattice: HangingPiece.safe_to_capture is true iff
at least one attacker could actually capture without exposing its own
king. extract_concepts's hanging_piece loop filters on it before flagging
anything - a piece nobody can safely take doesn't count, checked before
flagging rather than as an afterthought. (Required splitting find_hanging
into a raw scan collapse_criticality's own baseline depends on, plus the
enriched public version, to avoid the two calling each other forever.)

Persistence layer, deliberately shaped around "what can be quantified vs.
what can only be described": a new tactical_events table (chessdb/db.nu)
holds one row per individual finding with flat, graphable columns
(square, concept_name, side, severity, stage) next to a detail JSON column
carrying the full identifiable payload - no narrative/description column,
since synthesizing "why this matters" from the facts is interpretation,
not quantification, and belongs at read time (to whoever, or whichever
LLM, is actually explaining a position to a specific player) rather than
frozen into a row that could only serve one framing. New
chessdb::collapse-criticality plugin command exposes the Rust-only
primitive to the Nu side; new chess-tactical-events Nu command walks a
game's moves, extracts all five ladder concepts per position, and folds
the full collapse_criticality breakdown into hanging-piece detail.

Also bumped nu-plugin/nu-protocol 0.111 -> 0.114 to match the installed
nu CLI (no code changes needed, fully API-compatible) and verified the
whole pipeline live rather than just at the parse level: collapse-criticality
and hugm-eval both round-tripped exact matches against hand-derived
expectations over the real plugin protocol, and chess-tactical-events ran
against a real Scholar's Mate game imported through the actual
process-corpus pipeline - correctly found Black's f7 outnumbered and
White's own queen hanging at the exact position that makes the mate work,
idempotent on re-run.

Verified throughout: cargo test (48 lib + 28 motif, up from 43+24), cargo
clippy --all-targets zero warnings, STS 1499-position smoke test passes,
live nu-plugin round-trip confirmed for all three new/changed commands.

+++ nu_plugin_chessdb/src/process_corpus.rs
@@ -14,7 +14,7 @@ struct PendingPos {
     fen: String,
     hugm_score: i64,
     hugm_eval_arr: String,
-    state_id: u16,
+    state_id: u32,
     mate_in_1: i64,
     is_checkmate: i64,
 }
@@ -168,7 +168,7 @@ impl PluginCommand for ProcessCorpus {
                                 serde_json::to_string(&arr).unwrap_or_else(|_| "[]".to_string());
                             (rec.final_score, json_str, sid, mi1, cm)
                         }
-                        Err(_) => (0, "[]".to_string(), 0u16, 0, 0),
+                        Err(_) => (0, "[]".to_string(), 0u32, 0, 0),
                     };
                 PendingPos {
                     zobrist: fe.zobrist.clone(),
